requirements.txt
streamlit
facenet-pytorch
Pillow
torch
torchvision